Additional work on I2E Portfolio Service Layer.  Updated hibernate mapping for I2e Account roles due to nulls in composite key.

+++ src/gov/nih/nci/cbiit/scimgmt/entmaint/dao/I2ePortfolioDAO.java
@@ -163,7 +163,7 @@ public class I2ePortfolioDAO {
 	/**
 	 * get Audit Note by id
 	 */
-	public String getPortfolioNoteById(Integer id){
+	public String getPortfolioNoteById(Long id){
 		String note = "";
 		Criteria crit;
 		try {
